Add files via upload
- Google Coding Workshop Question (Spring 2017, University of Nevada, Reno)
- Description of program:
Retrieve license plate composed of English letters and digits from user. I retrieve the license plate data via terminal. Using the "re" (regular expression) python library, I am able to determine if the initial license plate entered by the user contains at least one letter which is a requirement for this question. The next objective is to extract the letters and store them in a list for further analysis. I then find the frequency of each letter found in the plate. This is accomplished by using the "Counter" method found in the "collections" python library. The resulting collection contains the letters found in the plate and treats them as the key, whereas the frequency of the letter itself is stored as the value. Using the “nltk” or Natural Learning Toolkit, I save the words found in nltk’s corpus in a list for letter frequency selection. For example, if a dictionary has the form